+++ apps/website/content/docs/changelog.md
@@ -2,6 +2,18 @@
 title: Changelog
 ---
 
+## v1.44.0 (2025-04-11)
+
+### ✨ New
+
+- feat: add JSX fragment factory to error message of `react-x/avoid-shorthand-fragment` by @Rel1cx in https://github.com/Rel1cx/eslint-react/pull/1051
+
+### 🪄 Improvements
+
+- perf: replace `valibot` with `@zod/mini` by @Rel1cx in https://github.com/Rel1cx/eslint-react/pull/1050
+
+**Full Changelog**: https://github.com/Rel1cx/eslint-react/compare/v1.43.0...v1.44.0
+
 ## v1.43.0 (2025-04-10)
 
 ### ✨ New